Create util package and move utilities inside
