+++ ippon/fight/permissions.py
@@ -1,7 +1,7 @@
 from rest_framework import permissions
 
 import ippon.fight.serializers as fs
-import ippon.permissions as ip
+import ippon.utils.permissions as ip
 import ippon.models.team_fight as tfm
 import ippon.models.tournament as tm
