+++ ippon/point/permissions.py
@@ -2,7 +2,7 @@ from rest_framework import permissions
 
 import ippon.models
 import ippon.models.fight
-import ippon.permissions as ip
+import ippon.utils.permissions as ip
 import ippon.point.serializers as pts
 import ippon.models.tournament as tm
