+++ ippon/tournament/permissions.py
@@ -3,7 +3,7 @@ import json
 from rest_framework import permissions
 
 import ippon.models.tournament as tm
-from ippon.permissions import is_user_admin_of_the_tournament
+from ippon.utils.permissions import is_user_admin_of_the_tournament
 
 
 class IsTournamentAdminOrReadOnlyTournament(permissions.BasePermission):
