+++ ippon/utils/debug.py
@@ -1,5 +1,2 @@
-BAD_PK = 0
-
-
 def print_comparison(cp1_json, data):
     print("++++++\n{}\n============\n{}----------".format(cp1_json, data))
