+++ ippon/utils/values.py
@@ -1,0 +1,1 @@
+BAD_PK = 0
